Fix BPcal per SPW plotting

+++ 13A-213/pipeline_scripts/casa_pipeline_lines.py
@@ -123,13 +123,20 @@ tb.open(ms_active + "/SPECTRAL_WINDOW")
 nspws = tb.getcol("NAME").shape[0]
 tb.close()
 
+# Final BP cal table now includes the stage number and step
+finalbpcal_name = glob(mySDM + '*.finalBPcal.b')
+if len(finalbpcal_name) == 0:
+    raise ValueError("Cannot find finalBPcal table name.")
+# Blindly assume we want the first name
+finalbpcal_name = finalbpcal_name[0]
+q
 for ii in range(nspws):
     filename = 'finalBPcal_amp_spw_' + str(ii) + '.png'
     syscommand = 'rm -rf ' + filename
     os.system(syscommand)
 
     default('plotcal')
-    caltable = mySDM + '.finalBPcal.b'
+    caltable = finalbpcal_name
     xaxis = 'freq'
     yaxis = 'amp'
     poln = ''
@@ -159,7 +166,7 @@ for ii in range(nspws):
     antPlot = str(ii * 3) + '~' + str(ii * 3 + 2)
 
     default('plotcal')
-    caltable = mySDM + '.finalBPcal.b'
+    caltable = finalbpcal_name
     xaxis = 'freq'
     yaxis = 'phase'
     poln = ''
